pyproject.toml

[project]
name = "vcenter-operator"
version = "0.1.0"
description = ""
authors = [
    {name = "Fabian Wiesel", email = "fabian.wiesel@sap.com"},
]
dependencies = [
    "attrs",
    "dnspython~=2.3.0",
    "masterpassword @ git+https://github.com/sapcc/masterpassword@main",
    "pyvmomi[sso]<9.0",
    "jinja2<4,>=3",
    "kubernetes~=28.0",
    "keystoneauth1",
    "dumb-init",
    "pyvmomi-extended @ git+https://github.com/sapcc/pyvmomi-extended@main"
    ]

requires-python = ">=3.10"
readme = "README.rst"
license = {text = "Apache-2.0"}

[project.scripts]
vcenter-operator = "vcenter_operator.cmd:main"

[build-system]
requires = ["setuptools ~= 65.5", "wheel"]

[tool.ruff]
# Allow lines to be as long as 120 characters.
line-length = 120
select = ["E", "F", "W", "I", "N", "UP"]

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = ruff
basepython = python3.10

[testenv:pep8]
commands = ruff
deps = ruff
basepython = python3.10
"""
